Renamed solve-wip.h -> solve.h - Now compiling test-solve-wip - but other tests not yet tested
