+++ linbox/algorithms/hybrid-det.h
@@ -254,13 +254,13 @@ namespace LinBox
 		PrimeIterator<IteratorCategories::HeuristicTag> genprime1(FieldTraits<mymodular>::bestBitSize(A.coldim()));
 
 		Integers ZZ;
-		RationalSolver < Integers , mymodular, PrimeIterator<IteratorCategories::HeuristicTag>, DixonTraits > RSolver(A. field(), genprime);
+		RationalSolver < Integers , mymodular, PrimeIterator<IteratorCategories::HeuristicTag>, Method::Dixon > RSolver(A. field(), genprime);
 #endif
-		RationalSolver < Integers , mymodular, PrimeIterator<IteratorCategories::HeuristicTag>, DixonTraits > RSolver;
+		RationalSolver < Integers , mymodular, PrimeIterator<IteratorCategories::HeuristicTag>, Method::Dixon > RSolver;
 
 		BlasVector<Integers> r_num1 (A.field(),A. coldim());
 
-		LastInvariantFactor < Integers ,RationalSolver < Integers, mymodular, PrimeIterator<IteratorCategories::HeuristicTag>, DixonTraits > >  LIF(RSolver);
+		LastInvariantFactor < Integers ,RationalSolver < Integers, mymodular, PrimeIterator<IteratorCategories::HeuristicTag>, Method::Dixon > >  LIF(RSolver);
 #ifdef _LB_H_DET_TIMING
 		BT.start();
 #endif
