+++ linbox/algorithms/lanczos.inl
@@ -249,7 +249,7 @@ namespace LinBox
 		}
 		else {
 			commentator().stop ("done", "Solve failed", "BlockLanczosSolver::solve");
-			throw SolveFailed ();
+			throw LinboxError ();
 		}
 	}
