+++ linbox/algorithms/mg-block-lanczos.h
@@ -37,7 +37,7 @@
 #include "linbox/field/archetype.h"
 #include "linbox/vector/vector-domain.h"
 #include "linbox/blackbox/archetype.h"
-#include "linbox/solutions/methods-wip.h"
+#include "linbox/solutions/methods.h"
 #include "linbox/matrix/dense-matrix.h"
 
 // I'm putting everything inside the LinBox namespace so that I can drop all of
@@ -74,7 +74,7 @@ namespace LinBox
 		 * @param traits @ref SolverTraits  structure describing user
 		 *               options for the solver
 		 */
-		MGBlockLanczosSolver (const Field &F, const MethodWIP::BlockLanczos &traits) :
+		MGBlockLanczosSolver (const Field &F, const Method::BlockLanczos &traits) :
 			_traits (traits), _field (&F), _VD (F), _MD (F), _randiter (F)
 			,_AV(F), _VTAV(F), _AVTAVSST_VTAV(F), _matT(F), _DEF(F)
             , _x(F), _y(F), _b(F), _tmp(F), _tmp1(F), _matM(F)
@@ -89,7 +89,7 @@ namespace LinBox
 		 *               options for the solver
 		 * @param r Random iterator to use for randomization
 		 */
-		MGBlockLanczosSolver (const Field &F, const MethodWIP::BlockLanczos &traits, typename Field::RandIter r) :
+		MGBlockLanczosSolver (const Field &F, const Method::BlockLanczos &traits, typename Field::RandIter r) :
 			_traits (traits), _field (&F), _VD (F), _MD (F), _randiter (r)
 			,_AV(F), _VTAV(F), _AVTAVSST_VTAV(F), _matT(F), _DEF(F)
             , _x(F), _y(F), _b(F), _tmp(F), _tmp1(F), _matM(F)
@@ -228,7 +228,7 @@ namespace LinBox
 
 		// Private variables
 
-		const MethodWIP::BlockLanczos _traits;
+		const Method::BlockLanczos _traits;
 		const Field              *_field;
 		VectorDomain<Field>       _VD;
 		MatrixDomain<Field>       _MD;
