+++ linbox/algorithms/rational-solver2.h
@@ -56,7 +56,7 @@ namespace LinBox
 	//template argument Field and RandomPrime are not used.
 	//Keep it just for interface consistency.
 	template <class Ring, class Field, class RandomPrime>
-	class RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits> {
+	class RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm> {
 
 	protected:
 		Ring r;
@@ -187,7 +187,7 @@ namespace LinBox
 	};
 #if __LINBOX_HAVE_CLAPACK
 	template <class Ring, class Field, class RandomPrime>
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::cblas_dgeinv(double* M, int n)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::cblas_dgeinv(double* M, int n)
 	{
 		enum CBLAS_ORDER order = CblasRowMajor;
 		int lda = n;
@@ -205,7 +205,7 @@ namespace LinBox
 	}
 
 	template <class Ring, class Field, class RandomPrime>
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::cblas_rsol (int n, const double* M, integer* numx, integer& denx, double* b)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::cblas_rsol (int n, const double* M, integer* numx, integer& denx, double* b)
 	{
 		if (n < 1) return 0;
 		double* IM = new double[n * n];
@@ -422,14 +422,14 @@ namespace LinBox
 
 	template <class Ring, class Field, class RandomPrime>
 	/* apply  y <- Ax */
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::cblas_dapply (int m, int n, const double* A, const double* x, double* y)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::cblas_dapply (int m, int n, const double* A, const double* x, double* y)
 	{
 		cblas_dgemv (CblasRowMajor, CblasNoTrans, m, n, 1, A, n, x, 1, 0, y, 1);
 		return 0;
 	}
 
 	template <class Ring, class Field, class RandomPrime>
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::cblas_mpzapply (int m, int n, const double* A, const integer* x, integer* y)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::cblas_mpzapply (int m, int n, const double* A, const integer* x, integer* y)
 	{
 		const double* p_A;
 		const integer* p_x;
@@ -449,7 +449,7 @@ namespace LinBox
 
 	template <class Ring, class Field, class RandomPrime>
 	template <class Elt>
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::printvec (const Elt* v, int n)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::printvec (const Elt* v, int n)
 	{
 		const Elt* p;
 		std::cout << '[';
@@ -461,7 +461,7 @@ namespace LinBox
 
 	template <class Ring, class Field, class RandomPrime>
 	//update num, *num <- *num * 2^shift + d
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::update_num (integer* num, int n, const double* d, int shift)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::update_num (integer* num, int n, const double* d, int shift)
 	{
 		integer* p_mpz;
 		integer tmp_mpz;
@@ -476,7 +476,7 @@ namespace LinBox
 
 	template <class Ring, class Field, class RandomPrime>
 	//update r = r * shift - M d
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::update_r_int (double* r, int n, const double* M, const double* d, int shift)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::update_r_int (double* r, int n, const double* M, const double* d, int shift)
 	{
 		double* p1;
 		const double* p2;
@@ -495,7 +495,7 @@ namespace LinBox
 
 	template <class Ring, class Field, class RandomPrime>
 	//update r = r * shift - M d
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::update_r_ll (double* r, int n, const double* M, const double* d, int shift)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::update_r_ll (double* r, int n, const double* M, const double* d, int shift)
 	{
 		double* p1;
 		const double* p2;
@@ -513,7 +513,7 @@ namespace LinBox
 	}
 
 	template <class Ring, class Field, class RandomPrime>
-	inline double RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::cblas_dOOnorm(const double* M, int m, int n)
+	inline double RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::cblas_dOOnorm(const double* M, int m, int n)
 	{
 		double norm = 0;
 		const double* p;
@@ -528,13 +528,13 @@ namespace LinBox
 	}
 
 	template <class Ring, class Field, class RandomPrime>
-	inline double RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::cblas_dmax (const int N, const double* a, const int inc)
+	inline double RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::cblas_dmax (const int N, const double* a, const int inc)
 	{
 		return fabs(a[cblas_idamax (N, a, inc)]);
 	}
 
 	template <class Ring, class Field, class RandomPrime>
-	inline int RationalSolver<Ring, Field, RandomPrime, NumSymNormTraits>::cblas_hbound (integer& b, int m, int n, const double* M)
+	inline int RationalSolver<Ring, Field, RandomPrime, Method::NumericSymbolicNorm>::cblas_hbound (integer& b, int m, int n, const double* M)
 	{
 		const  double* p;
 		integer tmp;
