+++ linbox/algorithms/wiedemann.h
@@ -72,7 +72,7 @@
 #include "linbox/matrix/sparse-matrix.h"
 #include "linbox/util/debug.h"
 #include "linbox/vector/vector-domain.h"
-#include "linbox/solutions/methods-wip.h"
+#include "linbox/solutions/methods.h"
 
 #include "linbox/algorithms/blackbox-container.h"
 #include "linbox/algorithms/blackbox-container-symmetric.h"
@@ -101,21 +101,21 @@ namespace LinBox
 
 			Squarize<Blackbox> B(&A);
 			BlackboxContainer<Field, Squarize<Blackbox> > TF (&B, A.field(), i);
-			MasseyDomain< Field, BlackboxContainer<Field, Squarize<Blackbox> > > WD (&TF, M.earlyTermThreshold ());
+			MasseyDomain< Field, BlackboxContainer<Field, Squarize<Blackbox> > > WD (&TF, M.earlyTerminationThreshold);
 
 			WD.minpoly (P, deg);
 		}
-		else if (M.symmetric ()) {
+		else if (M.shapeFlags & ShapeFlag::Symmetric) {
 			typedef BlackboxContainerSymmetric<Field, Blackbox> BBContainerSym;
 			BBContainerSym TF (&A, A.field(), i);
-			MasseyDomain< Field, BBContainerSym > WD (&TF, M.earlyTermThreshold ());
+			MasseyDomain< Field, BBContainerSym > WD (&TF, M.earlyTerminationThreshold);
 
 			WD.minpoly (P, deg);
 		}
 		else {
 			typedef BlackboxContainer<Field, Blackbox> BBContainer;
 			BBContainer TF (&A, A.field(), i);
-			MasseyDomain< Field, BBContainer > WD (&TF, M.earlyTermThreshold ());
+			MasseyDomain< Field, BBContainer > WD (&TF, M.earlyTerminationThreshold);
 
 			WD.minpoly (P, deg);
 #ifdef INCLUDE_TIMING
@@ -156,7 +156,7 @@ namespace LinBox
 			     Polynomial         &P,
 			     const Blackbox                            &A,
 			     const RingCategories::ModularTag          &tag,
-			     const Method::ExtensionWiedemann& M)
+			     const Method::WiedemannExtension& M)
 	{
 		typedef typename Blackbox::Field Field;
 		const Field& F = A.field();
@@ -206,7 +206,7 @@ namespace LinBox
 			     Polynomial         &P,
 			     const Blackbox                            &A,
 			     const RingCategories::ModularTag          &tag,
-			     const Method::ExtensionWiedemann& M)
+			     const Method::WiedemannExtension& M)
 	{
 		commentator().report (Commentator::LEVEL_ALWAYS,INTERNAL_WARNING) << " WARNING, no extension available, returning only a factor of the minpoly\n";
 		return minpoly(P, A, tag, Method::Wiedemann (M));
@@ -237,7 +237,7 @@ namespace LinBox
 		 * @param traits @ref SolverTraits  structure describing user
 		 *               options for the solver
 		 */
-		WiedemannSolver (const Field &F, const MethodWIP::Wiedemann &traits) :
+		WiedemannSolver (const Field &F, const Method::Wiedemann &traits) :
 			_traits (traits), _field (&F), _randiter (F), _VD (F)
 		{}
 
@@ -249,7 +249,7 @@ namespace LinBox
 		 * @param r Random iterator to use for randomization
 		 */
 		WiedemannSolver (const Field &F,
-				 const MethodWIP::Wiedemann &traits,
+				 const Method::Wiedemann &traits,
 				 typename Field::RandIter r) :
 			_traits (traits), _field (&F), _randiter (r), _VD (F)
 		{}
@@ -359,7 +359,7 @@ namespace LinBox
 		// Make an m x m lambda-sparse matrix, c.f. Mulders (2000)
 		SparseMatrix<Field> *makeLambdaSparseMatrix (size_t m);
 
-		MethodWIP::Wiedemann                      _traits;
+		Method::Wiedemann                      _traits;
 		const Field                         *_field;
 		typename Field::RandIter             _randiter;
 		VectorDomain<Field>                  _VD;
