+++ linbox/solutions/solve.h
@@ -1,7 +1,5 @@
-/* linbox/solutions/solve.h
+/*
  * Copyright(C) LinBox
- *  Evolved from an earlier one by Bradford Hovinen <hovinen@cis.udel.edu>
- *
  *
  * ========LICENCE========
  * This file is part of the library LinBox.
@@ -20,1091 +18,300 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  * ========LICENCE========
- *.
  */
 
 #ifndef __LINBOX_solve_H
 #define __LINBOX_solve_H
 
-/*! @file linbox/solutions/solve.h
- * @ingroup solutions
- * @brief System Solving.
- * @details NO DOC
- */
+#include <iostream> // @note Needed for givaro/ring-interface to compile
 
-#include <algorithm>
+#include <linbox/field/field-traits.h>
+#include <linbox/solutions/methods.h>
 
-#ifdef DEBUG_SOLVE
-#  define DEBUG_DIXON
-#  define DEBUG_INC
-#  define DEBUG_RR
-#  define DEBUG_LC
-#  define DEBUG_HB
-#  define RSTIMING
-#  define LIFTING_PROGRESS
-#endif
+namespace LinBox {
+    //
+    // solve
+    //
 
-// must fix this list...
-#include "linbox/algorithms/gauss.h"
-#include "linbox/algorithms/gauss-gf2.h"
-#include "linbox/algorithms/wiedemann.h"
-#include "linbox/algorithms/rational-solver.h"
-#include "linbox/algorithms/diophantine-solver.h"
-#include "linbox/matrix/factorized-matrix.h"
-#include "linbox/util/debug.h"
-#include "linbox/util/error.h"
-#include "linbox/vector/vector-domain.h"
-#include "linbox/solutions/methods.h"
-#include "linbox/algorithms/bbsolve.h"
-
-#ifdef __LINBOX_HAVE_MPI
-#include "linbox/algorithms/cra-distributed.h"
-#endif
-#include "linbox/algorithms/rational-cra2.h"
-
-#include "linbox/algorithms/varprec-cra-builder-early-multip.h"
-#include "linbox/algorithms/block-wiedemann.h"
-#include "linbox/algorithms/coppersmith.h"
-
-#ifdef __LINBOX_HAVE_IML
-#include "linbox/util/iml_wrapper.h"
-#endif
-
-namespace LinBox
-{
-
-        // for specialization with respect to the DomainCategory
-    template< class Vector, class Blackbox, class SolveMethod, class DomainCategory>
-    Vector & solve (Vector &            x,
-                    const Blackbox &                A,
-                    const Vector &          b,
-                    const DomainCategory &        tag,
-                    const SolveMethod &            M);
-
-        /** \brief Solve Ax = b, for x.
-         *
-         * Vector x such that Ax = b is returned.  In the case of a singular
-         * matrix A, if the system is consistent, a random solution is returned
-         * by default.  The method parameter may contain an indication that an
-         * arbitrary element of the solution space is acceptable, which can be
-         * faster to compute.  If the system is inconsistent the zero vector is
-         * returned.
-         *
-         * @warning no doc for when using what method
-         *
-         * @param [out] x solution
-         * @param [in]  A matrix
-         * @param [in]  b target
-         * @param [in]  M method to use (\see solutions/method.h)
-             * @return reference to \p x
-             */
-        // * \ingroup solutions
-        //and the SolveStatus, if non-null, is set to indicate inconsistency.
-    template< class Vector, class Blackbox, class SolveMethod>
-    Vector & solve (Vector &                x,
-                    const Blackbox &                A,
-                    const Vector &          b,
-                    const SolveMethod &             M)
+    /**
+     * \brief Solve Ax = b, for x.
+     *
+     * Returns a vector x such that Ax = b.
+     *
+     * Specifically:
+     * - A non singular: the unique solution is returned.
+     * - A singular:
+     *      - Consistent system: a random solution is returned.
+     *          The method parameter can contain an hint that an
+     *          arbitrary element of the solution space is acceptable instead,
+     *          which can be faster to compute if one doesn't expect a result in that case.
+     *      - Inconsistent system: LinboxMathInconsistentSystem is thrown.
+     *      - Internal failure: LinboxError is thrown.
+     *
+     * CategoryTag is defaulted to `FieldTraits<Matrix::Field>::categoryTag()` when omitted.
+     *
+     * - Method::Auto
+     *      - DenseMatrix   > Method::DenseElimination
+     *      - SparseMatrix  > Method::SparseElimination
+     *      - Otherwise
+     *      |   - Row or column dimension < LINBOX_USE_BLACKBOX_THRESHOLD > Method::Elimination
+     *      |   - Otherwise                                               > Method::Blackbox
+     * - Method::Elimination
+     *      - SparseMatrix  > Method::SparseElimination
+     *      - Otherwise     > Method::DenseElimination
+     * - Method::DenseElimination
+     *      - DenseMatrix
+     *      |   - ModularTag > `LQUPMatrix<Field>::left_solve`
+     *      |   - IntegerTag
+     *      |   |   - RatVector > Method::Dixon
+     *      |   |   - Otherwise > Error
+     *      |   - Otherwise > Error (Not implemented)
+     *      - Otherwise > Method::DenseElimination but copy to a DenseMatrix first
+     * - Method::SparseElimination
+     *      - SparseMatrix
+     *      |   - IntegerTag > Method::Dixon
+     *      |   - Otherwise > `GaussDomain<Field>::solveInPlace`
+     *      - Otherwise > Method::SparseElimination but copy to SparseMatrix first
+     * - Method::Cra
+     *      - IntegerTag
+     *      |   - Dispatch::Distributed > `RationalCraDistributed`
+     *      |   - Otherwise             > `RationalCra`
+     *      - Otherwise > Error
+     * - Method::Dixon
+     *      - IntegerTag
+     *      |   - DenseMatrix   > `RationalSolver<..., Method::Dixon>`
+     *      |   - SparseMatrix  > `RationalSolver<..., Method::SparseElimination>`
+     *      |   - Otherwise     >  Error
+     *      - Otherwise > Error
+     * - Method::Blackbox > Method::Wiedemann
+     * - Method::Wiedemann
+     *      - ModularTag > `WiedemannSolver`
+     *      - Otherwise  > Error
+     * - Method::BlockWiedemann [@deprecated, not tested]
+     *      - ModularTag > `BlockWiedemannSolver`
+     *      - Otherwise  > Error
+     * - Method::Coppersmith [@deprecated, not tested]
+     *      - ModularTag > `CoppersmithSolver`
+     *      - Otherwise  > Error
+     * - Method::Lanczos
+     *      - ModularTag > `LanczosSolver`
+     *      - Otherwise  > Error
+     * - Method::BlockLanczos
+     *      - ModularTag > `MGBlockLanczosSolver`
+     *      - Otherwise  > Error
+     * - Method::NumericSymbolicOverlap
+     *      - IntegerTag
+     *      |   - DenseMatrix > `RationalSolverSN<Ring, LPS<FMatrix>>`
+     *      |   - Otherwise   > Error
+     *      - Otherwise  > Error
+     * - Method::NumericSymbolicNorm
+     *      - IntegerTag > `RationalSolver<..., Method::NumericSymbolicNorm>`
+     *      - Otherwise  > Error
+     *
+     * @param [out] x solution, can be a rational solution (vector of numerators and one denominator)
+     * @param [in]  A matrix
+     * @param [in]  b target
+     * @param [in]  tag domain of computation
+     * @param [in]  m method to use (\see solutions/method.h)
+     * @return reference to \p x
+     */
+    template <class ResultVector, class Matrix, class Vector, class CategoryTag, class SolveMethod>
+    inline ResultVector& solve(ResultVector& x, const Matrix& A, const Vector& b, const CategoryTag& tag, const SolveMethod& m)
     {
-        return solve(x, A, b, typename FieldTraits<typename Blackbox::Field>::categoryTag(), M);
+        throw LinBoxError("solve<" + CategoryTag::name() + ", " + SolveMethod::name() + "> does not exists.");
     }
 
-        /*!
-         * the solve with default method.
-         */
-    template< class Vector, class Blackbox>
-    Vector& solve(Vector& x, const Blackbox& A, const Vector& b)
+    /**
+     * \brief Solve dispatcher for automated category tag.
+     */
+    template <class ResultVector, class Matrix, class Vector, class SolveMethod>
+    inline ResultVector& solve(ResultVector& x, const Matrix& A, const Vector& b, const SolveMethod& m)
+    {
+        return solve(x, A, b, typename FieldTraits<typename Matrix::Field>::categoryTag(), m);
+    }
+
+    /**
+     * \brief Solve dispatcher for automated solve method.
+     */
+    template <class ResultVector, class Matrix, class Vector>
+    inline ResultVector& solve(ResultVector& x, const Matrix& A, const Vector& b)
     {
         return solve(x, A, b, Method::Auto());
     }
 
-        // in methods.h FoobarMethod and Method::Foobar are the same class.
-        // in methods.h template<BB> bool useBB(const BB& A) is defined.
-
-        //! @internal specialize this on blackboxes which have local methods
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const Method::Auto& m)
+    /**
+     * \brief Solve specialisation on IntegerTag.
+     *
+     * This forward to the rational interface (num, den).
+     * But will only work if the ResultVector if a vector of some Rational type.
+     */
+    template <class ResultVector, class Matrix, class Vector, class SolveMethod>
+    typename std::enable_if<std::is_same<typename SolveMethod::CategoryTag, RingCategories::IntegerTag>::value,
+                            ResultVector&>::type
+    solve(ResultVector& x, const Matrix& A, const Vector& b, const RingCategories::IntegerTag& tag, const SolveMethod& m)
     {
-        if (useBB(A)) return solve(x, A, b, Method::Blackbox(m));
-        else return solve(x, A, b, Method::Elimination(m));
-    }
+        using Ring = typename Vector::Field;
+        using Element = typename Ring::Element;
 
-        /**  @internal Blackbox method specialisation */
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const Method::Blackbox& m)
-    {
-            // what is chosen here should be best and/or most reliable currently available choice
-            //      integer c; A.field().cardinality(c);
-            //      if (c < 100) return solve(x, A, b, Method::BlockLanczos(m));
-        return solve(x, A, b, Method::Wiedemann(m));
-    }
+        Vector xNum(b.field(), x.size());
+        Element xDen;
 
-        //x @todo temporary - fix this
-        //#define inBlasRange(p) true
+        solve(xNum, xDen, A, b, tag, m);
 
-        /**  @internal Elimination method specialisation */
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const Method::Elimination& m)
-    {
-        integer c, p;
-        A.field().cardinality(c);
-        A.field().characteristic(p);
-        return solve(x, A, b,
-                     typename FieldTraits<typename BB::Field>::categoryTag(),
-                     Method::DenseElimination(m));
-    }
-
-        //! @internal inplace Sparse Elimination.
-    template <class Vector, class Field>
-    Vector& solveInPlace(Vector& x, SparseMatrix<Field, SparseMatrixFormat::SparseSeq>& A, const Vector& b, const Method::SparseElimination& m)
-    {
-        commentator().start ("Sparse Elimination Solve In Place", "sesolveInPlace");
-        GaussDomain<Field> GD ( A.field() );
-        GD.solveInPlace(x, A, b);
-        commentator().stop ("done", NULL, "sesolveInPlace");
-        return x;
-    }
-
-    template <class Vector, class Field, class Random>
-    Vector& solveInPlace(Vector& x, SparseMatrix<Field, SparseMatrixFormat::SparseSeq>& A, const Vector& b, const Method::SparseElimination& m, Random& generator)
-    {
-        commentator().start ("Sparse Elimination Solve In Place with random solution", "sesolveInPlace");
-        GaussDomain<Field> GD ( A.field() );
-        GD.solveInPlace(x, A, b, generator);
-        commentator().stop ("done", NULL, "sesolveInPlace");
-        return x;
-    }
-
-
-        //! @internal  Change of representation to be able to call the sparse elimination
-    template <class Vector, class Blackbox>
-    Vector& solve(Vector& x, const Blackbox& A, const Vector& b,
-                  const Method::SparseElimination& m)
-    {
-        typedef typename Blackbox::Field Field;
-        typedef SparseMatrix<Field,SparseMatrixFormat::SparseSeq> SparseBB;
-        SparseBB SpA(A.field(), A.rowdim(), A.coldim());
-        MatrixHom::map(SpA, A);
-        return solveInPlace(x, SpA, b, m);
-    }
-
-    template <class Vector, class Blackbox, class Random>
-    Vector& solve(Vector& x, const Blackbox& A, const Vector& b,
-                  const Method::SparseElimination& m, Random& generator)
-    {
-        typedef typename Blackbox::Field Field;
-        typedef SparseMatrix<Field,SparseMatrixFormat::SparseSeq> SparseBB;
-        SparseBB SpA(A.field(), A.rowdim(), A.coldim());
-        MatrixHom::map(SpA, A);
-        return solveInPlace(x, SpA, b, generator);
-    }
-
-        //! @internal specialisation for inplace SparseElimination on GF2
-    template <class Vector>
-    Vector& solveInPlace(Vector& x,
-                    GaussDomain<GF2>::Matrix    &A,
-                    const Vector& b,
-                    const Method::SparseElimination& m)
-    {
-        commentator().start ("Sparse Elimination Solve In Place over GF2", "GF2sesolveInPlace");
-        GaussDomain<GF2> GD ( A.field() );
-        GD.solveInPlace(x, A, b);
-        commentator().stop ("done", NULL, "GF2sesolveInPlace");
-        return x;
-    }
-    template <class Vector, class Random>
-    Vector& solveInPlace(Vector& x,
-                    GaussDomain<GF2>::Matrix    &A,
-                    const Vector& b,
-                    const Method::SparseElimination& m,
-                    Random& generator)
-    {
-        commentator().start ("Sparse Elimination Solve In Place over GF2", "GF2sesolveInPlace");
-        GaussDomain<GF2> GD ( A.field() );
-        GD.solveInPlace(x, A, b, generator);
-        commentator().stop ("done", NULL, "GF2sesolveInPlace");
-        return x;
-    }
-
-        //! @internal specialisation for SparseElimination on GF2
-    template <class Vector>
-    Vector& solve(Vector& x,
-                  GaussDomain<GF2>::Matrix    &A,
-                  const Vector& b,
-                  const Method::SparseElimination& m)
-    {
-            // We make a copy
-        GaussDomain<GF2>::Matrix SpA(A.field(), A.rowdim(), A.coldim());
-        MatrixHom::map(SpA, A );
-        return solveInPlace(x, SpA, b, m);
-    }
-    template <class Vector, class Random>
-    Vector& solve(Vector& x,
-                  GaussDomain<GF2>::Matrix    &A,
-                  const Vector& b,
-                  const Method::SparseElimination& m,
-                  Random& generator)
-    {
-            // We make a copy
-        GaussDomain<GF2>::Matrix SpA(A.field(), A.rowdim(), A.coldim());
-        MatrixHom::map(SpA, A );
-        return solveInPlace(x, SpA, b, m, generator);
-    }
-
-        //! @internal Generic Elimination for SparseMatrix
-    template <class Vector, class Field>
-    Vector& solve(Vector& x, const SparseMatrix<Field>& A, const Vector& b,
-                  const Method::Elimination& m)
-    {
-            //             bool consistent = false;
-            // sparse elimination based solver can be called here ?
-            // For now we call the dense one
-
-        return solve(x, A, b,
-                     typename FieldTraits<typename SparseMatrix<Field>::Field>::categoryTag(),
-                     Method::DenseElimination(m));
-
-#if 0
-        if ( ! consistent ) {  // we will return the zero vector
-            for (typename Vector::iterator i = x.begin(); i != x.end(); ++i) *i = A.field().zero;
-        }
-        return x;
-#endif
-    }
-        // DenseElimination section ///////////////////
-
-        //! @internal Generic Elimination on Z/pZ (convert A to DenseMatrix)
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const RingCategories::ModularTag & tag,
-                  const Method::DenseElimination& m)
-    {
-        BlasMatrix<typename BB::Field> B(A); // copy A into a BlasMatrix
-        return solve(x, B, b, tag, m);
-    }
-
-        //! @internal Generic Elimination for DenseMatrix on Z/pZ
-    template <class Vector, class Field>
-    Vector& solve(Vector& x, const BlasMatrix<Field>& A, const Vector& b,
-                  const RingCategories::ModularTag & tag,
-                  const Method::DenseElimination& m)
-    {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-
-        commentator().start ("Solving linear system (FFLAS LQUP)", "LQUP::left_solve");
-            //bool consistent = false;
-        LQUPMatrix<Field> LQUP(A);
-            //FactorizedMatrix<Field> LQUP(A);
-
-        LQUP.left_solve(x, b);
-
-#if 0
-            // this should be implemented directly in left_solve
-        if ( ! consistent ) {  // we will return the zero vector
-            for (typename Vector::iterator i = x.begin(); i != x.end(); ++i)
-                *i = A.field().zero;
-        }
-#endif
-        commentator().stop ("done", NULL, "LQUP::left_solve");
-
-        return x;
-    }
-
-    template <class Vector, class Field>
-    Vector& solve(Vector& x, const BlasMatrix<Field>& A, const Vector& b,
-                  const RingCategories::ModularTag & tag,
-                  const Method::Dixon & m)
-    {
-        throw LinBoxFailure("You cannot do this");
-    }
-
-        /* Integer tag Specialization for Dixon method:
-         * 2 interfaces:
-         *   - the output is a common denominator and a vector of numerator (no need of rational number)
-         *   - the output is a vector of rational
-         */
-
-
-        // error handler for bad use of the integer solver API
-        //! @internal Generic Elimination for  Integer matrices
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const RingCategories::IntegerTag & tag,
-                  const Method::DenseElimination& m)
-    {
-        std::cout<<"try to solve system over the integer\n"
-                 <<"the API need either \n"
-                 <<" - a vector of rational as the solution \n"
-                 <<" - or an integer for the common denominator and a vector of integer for the numerators\n\n";
-        throw LinboxError("bad use of integer API solver\n");
-
-    }
-
-#if 0
-    template <class RatVector, class Vector, class BB, class MethodTraits>
-    Vector& solve(RatVector& x, const BB& A, const Vector& b,
-                  const RingCategories::RationalTag & tag,
-                  const MethodTraits& m)
-    {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-
-        commentator().start ("Rational CRA Solve", "Rsolve");
-        typedef Givaro::Modular<double> Field;
-        PrimeIterator<IteratorCategories::HeuristicTag> genprime(FieldTraits<Field>::bestBitSize(A.coldim()));
-
-        RationalRemainder2< VarprecCraBuilderEarlyMultip<Field> > rra(3UL);//using default RR method
-        IntegerModularSolve<BB,Vector,MethodTraits > iteration(A, b, m);
-        integer den;
-        BlasVector<Givaro::ZRing<Integer> > num(A.field(),A.coldim());
-
-        rra(num, den, iteration, genprime);
-
-        typename RatVector::iterator it_x= x.begin();
-        typename BlasVector<Givaro::ZRing<Integer> >::const_iterator it_num= num.begin();
-
-        for (; it_x != x.end(); ++it_x, ++it_num){
-            integer g = gcd( *it_num, den);
-            *it_x = typename RatVector::value_type(*it_num/g, den/g);
+        // The denominator being zero means computation failure
+        if (b.field().isZero(xDen)) {
+            throw LinboxError("Rational solve failed.");
         }
 
-        commentator().stop ("done", NULL, "Rsolve");
-        return x;
-    }
-#endif
-        // error handler for non defined solver over rational domain
-#if 0
-    template <class Vector, class BB, class MethodTraits>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const RingCategories::RationalTag & tag,
-                  const MethodTraits& m)
-    {
-        throw LinboxError("LinBox ERROR: solver not yet defined over rational domain");
-    }
-#endif
-
-        /*
-         * 1st integer solver API :
-         * solution is a vector of rational numbers
-         * RatVector is assumed to be the type of a vector of rational number
-        */
-
-        // default API (method is DenseElimination)
-    template<class RatVector, class Vector, class BB>
-    RatVector& solve(RatVector& x, const BB &A, const Vector &b)
-    {
-        return solve(x, A, b, Method::DenseElimination());
-    }
-
-        // API with Auto method
-    template<class RatVector, class Vector, class BB>
-    RatVector& solve(RatVector& x, const BB &A, const Vector &b,
-                     const Method::Auto &m)
-    {
-        if (useBB(A))
-            return solve(x, A, b, Method::Blackbox(m));
-        else
-            return solve(x, A, b, Method::Elimination(m));
-    }
-
-        // API with Blackbox method
-    template<class RatVector, class Vector, class BB>
-    RatVector& solve(RatVector& x, const BB &A, const Vector &b, const Method::Blackbox &m)
-    {
-        return solve(x, A, b, Method::Wiedemann(m));
-    }
-
-        // API with Elimination method
-    template<class RatVector, class Vector, class BB>
-    RatVector& solve(RatVector& x, const BB &A, const Vector &b, const Method::Elimination &m)
-    {
-        return solve(x, A, b,  Method::DenseElimination(m));
-    }
-
-
-        // launcher of specialized solver depending on the MethodTrait
-    template<class RatVector, class Vector, class BB, class MethodTraits>
-    RatVector& solve(RatVector& x, const BB &A, const Vector &b, const MethodTraits &m)
-    {
-        return solve(x, A, b, typename FieldTraits<typename BB::Field>::categoryTag(),  m);
-    }
-
-
-        /* Specializations for DenseElimination over the integers
-         */
-
-        // input matrix is generic (copying it into a BlasMatrix)
-    template <class RatVector, class Vector, class BB>
-    RatVector& solve(RatVector& x, const BB& A, const Vector& b,
-                     const RingCategories::IntegerTag & tag,
-                     const Method::DenseElimination& m)
-    {
-        BlasMatrix<typename BB::Field> B(A); // copy A into a BlasMatrix
-        return solve(x, B, b, tag, m);
-    }
-
-        // input matrix is a BlasMatrix (no copy)
-    template <class RatVector, class Vector, class Ring>
-    RatVector& solve(RatVector& x, const BlasMatrix<Ring>& A, const Vector& b,
-                     const RingCategories::IntegerTag & tag,
-                     const Method::DenseElimination& m)
-    {
-
-        Method::Dixon mDixon(m);
-        typename Ring::Element d;
-        BlasVector<Ring> num(A.field(),A.coldim());
-        solve (num, d, A, b, tag, mDixon);
-
-        typename RatVector::iterator it_x= x.begin();
-        typename BlasVector<Ring>::const_iterator it_num= num.begin();
-        integer n,den;
-        A.field().convert(den,d);
-        for (; it_x != x.end(); ++it_x, ++it_num){
-            A.field().convert(n, *it_num);
-            *it_x = typename RatVector::value_type(n, den);
+        // Copy result back to ResultVector
+        auto iXNum = xNum.begin();
+        for (auto iX = x.begin(); iX != x.end(); ++iX) {
+            *iX = typename ResultVector::value_type(*iXNum, xDen);
+            ++iXNum;
         }
 
         return x;
     }
 
-        /*!
-         * 2nd integer solver API :
-         * solution is a formed by a common denominator and a vector of integer numerator
-         * solution is num/d
-        * BB: why not a struct RatVector2 { IntVector _n ; Int _d } ; ?
-        */
-
-        //@{
-
-        // default API (method is DenseElimination)
-    template< class Vector, class BB>
-    Vector& solve(Vector &x, typename BB::Field::Element &d, const BB &A, const Vector &b)
+    /**
+     * \brief Rational solve Ax = b, for x expressed as xNum/xDen.
+     *
+     * Second interface for solving, only valid for RingCategories::IntegerTag.
+     *
+     * Solve with this interface will usually go for CRA or Dixon lifting,
+     * as non-modular elimination would snowball elements to very big values.
+     */
+    template <class Matrix, class Vector, class CategoryTag, class SolveMethod>
+    inline void solve(Vector& xNum, typename Vector::Field::Element& xDen, const Matrix& A, const Vector& b,
+                      const CategoryTag& tag, const SolveMethod& m)
     {
-        return solve(x, d, A, b, typename FieldTraits<typename BB::Field>::categoryTag(),  Method::DenseElimination());
+        throw LinBoxError("Rational solve is only valid for RingCategories::IntegerTag.");
     }
 
-        // launcher of specialized solver depending on the MethodTraits
-    template< class Vector, class BB, class MethodTraits>
-    Vector& solve(Vector &x, typename BB::Field::Element &d, const BB &A, const Vector &b, const MethodTraits &m)
+    /**
+     * \brief Rational solve dispatcher for unimplemented methods.
+     */
+    template <class Matrix, class Vector, class SolveMethod>
+    inline void solve(Vector& xNum, typename Vector::Field::Element& xDen, const Matrix& A, const Vector& b,
+                      const RingCategories::IntegerTag& tag, const SolveMethod& m)
     {
-        return solve(x, d, A, b, typename FieldTraits<typename BB::Field>::categoryTag(), m);
+        throw LinBoxError("Rational solve<RingCategories::IntegerTag, " + SolveMethod::name() + "> does not exists.");
     }
 
-        /* Specialization for DenseElimination over the integers
-         */
-
-        // input matrix is generic (copying it into a BlasMatrix)
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, typename BB::Field::Element &d, const BB& A, const Vector& b,
-                  const RingCategories::IntegerTag & tag,
-                  const Method::DenseElimination& m)
+    /**
+     * \brief Rational solve dispatcher for automated category tag.
+     */
+    template <class Matrix, class Vector, class SolveMethod>
+    inline void solve(Vector& xNum, typename Vector::Field::Element& xDen, const Matrix& A, const Vector& b, const SolveMethod& m)
     {
-        BlasMatrix<typename BB::Field> B(A); // copy A into a BlasMatrix
-        return solve(x, d, B, b, tag, m);
+        solve(xNum, xDen, A, b, typename FieldTraits<typename Matrix::Field>::categoryTag(), m);
     }
 
-        // input matrix is a BlasMatrix (no copy)
-    template <class Vector, class Ring>
-    Vector& solve(Vector& x, typename Ring::Element &d,
-                  const BlasMatrix<Ring>& A, const Vector& b,
-                  const RingCategories::IntegerTag & tag,
-                  const Method::DenseElimination& m)
+    /**
+     * \brief Rational solve dispatcher for automated solve method.
+     */
+    template <class Matrix, class Vector>
+    inline void solve(Vector& xNum, typename Vector::Field::Element& xDen, const Matrix& A, const Vector& b)
     {
-            //!@bug check we don't copy
-        Method::Dixon mDixon(m);
-        return solve(x, d, A, b, tag, mDixon);
+        solve(xNum, xDen, A, b, Method::Auto());
     }
 
-        // input matrix is a SparseMatrix (no copy)
-    template <class Vect, class Ring>
-    Vect& solve(Vect& x, typename Ring::Element &d,
-                const SparseMatrix<Ring, SparseMatrixFormat::SparseSeq>& A,
-                const Vect& b,
-                const RingCategories::IntegerTag & tag,
-                const Method::SparseElimination& m)
+    //
+    // solveInPlace
+    //
+
+    /**
+     * \brief Solve Ax = b, for x.
+     *
+     * Returns a vector x such that Ax = b.
+     * A can be modified.
+     *
+     * See documentation for `solve`.
+     */
+    template <class ResultVector, class Matrix, class Vector, class CategoryTag, class SolveMethod>
+    inline ResultVector& solveInPlace(ResultVector& x, Matrix& A, const Vector& b, const CategoryTag& tag, const SolveMethod& m)
     {
-        Method::Dixon mDixon(m);
-        return solve(x, d, A, b, tag, mDixon);
+        // @note This is called if the specialization has not been implemented,
+        // which means there might not be any "in place" version of the solve.
+        // So, we just forward to classic solve.
+
+        return solve(x, A, b, tag, m);
     }
 
-
-
-        /** \brief solver specialization with the 2nd API and DixonTraits over integer (no copying)
-         */
-    template <class Vector, class Ring>
-    Vector& solve(Vector& x, typename Ring::Element &d,
-                  const BlasMatrix<Ring>& A,
-                  const Vector& b,
-                  const RingCategories::IntegerTag tag, Method::Dixon& m)
+    /**
+     * \brief Solve in place dispatcher for automated category tag.
+     */
+    template <class ResultVector, class Matrix, class Vector, class SolveMethod>
+    inline ResultVector& solveInPlace(ResultVector& x, Matrix& A, const Vector& b, const SolveMethod& m)
     {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-
-        commentator().start ("Padic Integer Blas-based Solving ", "solving");
-
-        typedef Givaro::Modular<double> Field;
-        PrimeIterator<IteratorCategories::HeuristicTag> genprime(FieldTraits<Field>::bestBitSize(A.coldim()));
-        RationalSolver<Ring, Field, PrimeIterator<IteratorCategories::HeuristicTag>, DixonTraits> rsolve(A.field(), genprime);
-        SolverReturnStatus status = SS_OK;
-
-            // if singularity unknown and matrix is square, we try nonsingular solver
-        switch ( m.singular() ) {
-            case Specifier::SINGULARITY_UNKNOWN:
-                switch (A.rowdim() == A.coldim() ?
-                        status=rsolve.solveNonsingular(x, d, A, b, false ,(int)m.maxTries()) : SS_SINGULAR) {
-                    case SS_OK:
-                        m.singular(Specifier::NONSINGULAR);
-                        break;
-                    case SS_SINGULAR:
-                        switch (m.solution()){
-                            case DixonTraits::DETERMINIST:
-                                status= rsolve.monolithicSolve(x, d, A, b, false, false, (int)m.maxTries(),
-                                                               (m.certificate()? SL_LASVEGAS: SL_MONTECARLO));
-                                break;
-                            case DixonTraits::RANDOM:
-                                status= rsolve.monolithicSolve(x, d, A, b, false, true, (int)m.maxTries(),
-                                                               (m.certificate()? SL_LASVEGAS: SL_MONTECARLO));
-                                break;
-                            case DixonTraits::DIOPHANTINE:
-                            {
-                                DiophantineSolver<RationalSolver<Ring,Field,PrimeIterator<IteratorCategories::HeuristicTag>, DixonTraits> > dsolve(rsolve);
-                                status= dsolve.diophantineSolve(x, d, A, b, (int)m.maxTries(),
-                                                                (m.certificate()? SL_LASVEGAS: SL_MONTECARLO));
-                            }
-                            break;
-                            default:
-                                break;
-                        }
-                        break;
-                    default:
-                        break;
-                }
-                break;
-
-            case Specifier::NONSINGULAR:
-                rsolve.solveNonsingular(x, d, A, b, false ,(int)m.maxTries());
-                break;
-
-            case Specifier::SINGULAR:
-                switch (m.solution()){
-                    case DixonTraits::DETERMINIST:
-                        status= rsolve.monolithicSolve(x, d, A, b,
-                                                       false, false, (int)m.maxTries(),
-                                                       (m.certificate()? SL_LASVEGAS: SL_MONTECARLO));
-                        break;
-
-                    case DixonTraits::RANDOM:
-                        status= rsolve.monolithicSolve(x, d, A, b,
-                                                       false, true, (int)m.maxTries(),
-                                                       (m.certificate()? SL_LASVEGAS: SL_MONTECARLO));
-                        break;
-
-                    case DixonTraits::DIOPHANTINE:
-                    {
-                        DiophantineSolver<RationalSolver<Ring,Field,PrimeIterator<IteratorCategories::HeuristicTag>, DixonTraits> > dsolve(rsolve);
-                        status= dsolve.diophantineSolve(x, d, A, b, (int)m.maxTries(),
-                                                        (m.certificate()? SL_LASVEGAS: SL_MONTECARLO));
-                    }
-                    break;
-
-                        //default:
-                        //  break;
-                }
-            default:
-                break;
-        }
-
-        commentator().stop("done", NULL, "solving");
-
-        if ( status == SS_INCONSISTENT ) {
-            throw LinboxMathInconsistentSystem("Linear system is inconsistent");
-                //          for (typename Vector::iterator i = x.begin(); i != x.end(); ++i) *i = A.field().zero;
-        }
-#ifdef RSTIMING
-        rsolve.reportTimes(std::clog);
-#endif
-        return x;
+        return solveInPlace(x, A, b, typename FieldTraits<typename Matrix::Field>::categoryTag(), m);
     }
 
-        /** \brief solver specialization with the 2nd API and DixonTraits over integer (no copying)
-         */
-    template <class Vect, class Ring>
-    Vect& solve(Vect& x, typename Ring::Element &d,
-                const SparseMatrix<Ring, SparseMatrixFormat::SparseSeq> & A,
-                const Vect& b,
-                const RingCategories::IntegerTag tag,
-                Method::Dixon& m)
+    /**
+     * \brief Solve in place dispatcher for automated solve method.
+     */
+    template <class ResultVector, class Matrix, class Vector>
+    inline ResultVector& solveInPlace(ResultVector& x, Matrix& A, const Vector& b)
     {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-
-        commentator().start ("Padic Integer Sparse Elimination Solving", "solving");
-
-        typedef Givaro::Modular<double> Field;
-        PrimeIterator<IteratorCategories::HeuristicTag> genprime(FieldTraits<Field>::bestBitSize(A.coldim()));
-        RationalSolver<Ring, Field, PrimeIterator<IteratorCategories::HeuristicTag>, SparseEliminationTraits> rsolve(A.field(), genprime);
-        SolverReturnStatus status = SS_OK;
-        status=rsolve.solve(x, d, A, b,(int)m.maxTries());
-
-        commentator().stop("done", NULL, "solving");
-
-        if ( status == SS_INCONSISTENT ) {
-            throw LinboxMathInconsistentSystem("Linear system is inconsistent");
-                //          for (typename Vect::iterator i = x.begin(); i != x.end(); ++i) *i = A.field().zero;
-        }
-        return x;
+        return solveInPlace(x, A, b, Method::Auto());
     }
 
-        //@}
-
-        // Lanczos ////////////////
-        // may throw SolverFailed or InconsistentSystem
-
-        // Wiedemann section ////////////////
-
-        // may throw SolverFailed or InconsistentSystem
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const RingCategories::ModularTag & tag,
-                  const Method::Wiedemann& m)
+    /**
+     * \brief Rational solve in place Ax = b, for x expressed as xNum/xDen.
+     * The matrix A might be modified.
+     *
+     * Second interface for solving in place, only valid for RingCategories::IntegerTag.
+     */
+    template <class Matrix, class Vector, class SolveMethod, class CategoryTag>
+    inline void solveInPlace(Vector& xNum, typename Vector::Field::Element& xDen, Matrix& A, const Vector& b,
+                             const CategoryTag& tag, const SolveMethod& m)
     {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
+        // @note This is called if the specialization has not been implemented,
+        // which means there might not be any "in place" version of the solve.
+        // So, we just forward to classic solve.
 
-            // adapt to earlier signature of wiedemann solver
-        solve(A, x, b, A.field(), m);
-        return x;
+        solve(xNum, xDen, A, b, tag, m);
     }
 
-        // Only for nonsingular system for now.
-        // may throw SolverFailed or InconsistentSystem
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const RingCategories::ModularTag & tag,
-                  const Method::BlockWiedemann& m)
+    /**
+     * \brief Rational solve in place dispatcher for automated category tag.
+     */
+    template <class Matrix, class Vector, class SolveMethod>
+    inline void solveInPlace(Vector& xNum, typename Vector::Field::Element& xDen, Matrix& A, const Vector& b,
+                             const SolveMethod& m)
     {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-
-            // adapt to earlier signature of wiedemann solver
-        typedef BlasMatrixDomain<typename BB::Field> Context;
-        Context BMD(A.field());
-        BlockWiedemannSolver<Context> BWS(BMD,m.blockingFactor(), m.blockingFactor()+1);
-            //BWS.solveNonSingular(x, A, b);
-        BWS.solve(x, A, b);
-        return x;
+        solveInPlace(xNum, xDen, A, b, typename FieldTraits<typename Matrix::Field>::categoryTag(), m);
     }
 
-        // Only for nonsingular system for now.
-        // may throw SolverFailed or InconsistentSystem
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, const BB& A, const Vector& b,
-                  const RingCategories::ModularTag & tag,
-                  const Method::Coppersmith& m)
+    /**
+     * \brief Rational solve in place dispatcher for automated solve method.
+     */
+    template <class Matrix, class Vector>
+    inline void solveInPlace(Vector& xNum, typename Vector::Field::Element& xDen, Matrix& A, const Vector& b)
     {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-
-            // adapt to earlier signature of wiedemann solver
-        CoppersmithSolver<typename BB::Field> cs(A.field());
-        cs.solveNonsingular(x, A, b);
-        return x;
-    }
-
-        /* remark 1.  I used copy constructors when switching method types.
-           But if the method types are (empty) child classes of a common  parent class containing
-           all the information, then casts can be used in place of copies.
-        */
-
-} // LinBox
-
-
-
-#include "linbox/ring/modular.h"
-#include "linbox/algorithms/rational-cra.h"
-#include "linbox/algorithms/rational-cra-builder-early-multip.h"
-#include "linbox/randiter/random-prime.h"
-#include "linbox/algorithms/matrix-hom.h"
-#include "linbox/vector/vector-traits.h"
-
-namespace LinBox
-{ /*  Integer */
-
-
-    template <class Blackbox, class Vector, class MyMethod>
-    struct IntegerModularSolve {
-        const Blackbox &A;
-        const Vector &B;
-        const MyMethod &M;
-
-        IntegerModularSolve(const Blackbox& b, const Vector& v, const MyMethod& n) :
-                A(b), B(v), M(n)
-            {}
-
-
-        template<typename Field>
-        typename Rebind<Vector, Field>::other& operator()(typename Rebind<Vector, Field>::other& x, const Field& F) const
-            {
-                typedef typename Blackbox::template rebind<Field>::other FBlackbox;
-                FBlackbox Ap(A, F);
-
-                typedef typename Rebind<Vector, Field>::other FVector;
-                FVector Bp(F, B);
-
-                VectorWrapper::ensureDim (x, A.coldim());
-                return solve( x, Ap, Bp, M);
-            }
-    };
-
-
-        //BB: How come I have to change the name so it works when directly called ?
-    template <class Vector, class BB, class MyMethod>
-    Vector& solveCRA(Vector& x, typename BB::Field::Element& d, const BB& A, const Vector& b,
-                     const RingCategories::IntegerTag & tag,
-                     const MyMethod& M
-#ifdef __LINBOX_HAVE_MPI
-                     ,Communicator   *C = NULL
-#endif
-                     )
-    {
-#ifdef __LINBOX_HAVE_MPI    //MPI parallel version
-
-        Integer den(1);
-        if(!C || C->rank() == 0){
-            if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-                throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-            commentator().start ("Integer CRA Solve", "Isolve");
-        }
-
-        // 0.7213475205 is an upper approximation of 1/(2log(2))
-        PrimeIterator<IteratorCategories::HeuristicTag> genprime((unsigned int)( 26 -(int)ceil(log((double)A.rowdim())*0.7213475205)));
-
-        BlasVector<Givaro::ZRing<Integer>> num(A.field(),A.coldim());
-
-        IntegerModularSolve<BB,Vector,MyMethod> iteration(A, b, M);
-        RationalCraDistributed< RationalCraBuilderEarlyMultip< Givaro::Modular<double> > > mpicra(3UL, C);
-
-        mpicra(num, den, iteration, genprime);
-
-        if(!C || C->rank() == 0){
-            typename Vector::iterator it_x= x.begin();
-            typename BlasVector<Givaro::ZRing<Integer>>::const_iterator it_num= num.begin();
-
-                // convert the result
-            for (; it_x != x.end(); ++it_x, ++it_num)
-                A.field().init(*it_x, *it_num);
-
-            A.field().init(d, den);
-
-            commentator().stop ("done", NULL, "Isolve");
-            return x;
-        }
-#else   //serial version
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-        commentator().start ("Integer CRA Solve", "Isolve");
-
-
-        // 0.7213475205 is an upper approximation of 1/(2log(2))
-        PrimeIterator<IteratorCategories::HeuristicTag> genprime((unsigned int)( 26 -(int)ceil(log((double)A.rowdim())*0.7213475205)));
-            //         RationalCra< Givaro::Modular<double> > rra((double)
-            //                                                  ( A.coldim()/2.0*log((double) A.coldim()) ) );
-
-            // use of integer due to non genericity of rra (PG 2005-09-01)
-        Integer den(1);
-        BlasVector<Givaro::ZRing<Integer>> num(A.field(),A.coldim());
-
-        IntegerModularSolve<BB,Vector,MyMethod> iteration(A, b, M);
-        RationalCra< RationalCraBuilderEarlyMultip< Givaro::Modular<double> > > rra(3UL);
-        rra(num, den, iteration, genprime); //rra(x, d, iteration, genprime);
-        typename Vector::iterator it_x= x.begin();
-        typename BlasVector<Givaro::ZRing<Integer>>::const_iterator it_num= num.begin();
-            // convert the result
-        for (; it_x != x.end(); ++it_x, ++it_num)
-            A.field().init(*it_x, *it_num);
-        A.field().init(d, den);
-        commentator().stop ("done", NULL, "Isolve");
-        return x;
-#endif
-    }
-
-
-
-
-
-        //BB: How come SparseElimination needs this ?
-        // may throw SolverFailed or InconsistentSystem
-    template <class Vector, class BB, class MyMethod>
-    Vector& solve(Vector& x, typename BB::Field::Element& d, const BB& A, const Vector& b,
-                  const RingCategories::IntegerTag & tag,
-                  const MyMethod& M)
-    {
-        Method::Dixon mDixon(M);
-        return solve(x,d,A,b,tag,mDixon);
-            //return solveCRA(x,d,A,b,tag,M);
-    }
-
-
-#if 0 /*  not working */
-    template <class Vector, class Field>
-    Vector& solve(Vector& x, typename Field::Element& d, const BlasMatrix<Field>& A, const Vector& b,
-                  const RingCategories::IntegerTag & tag,
-                  const Method::CRA & M)
-    {
-        return solve(x,d,A,b,tag,M.iterationMethod());
-    }
-
-    template <class Vector, class BB>
-    Vector& solve(Vector& x, typename BB::Field::Element& d, const BB& A, const Vector& b,
-                  const RingCategories::ModularTag & tag,
-                  const Method::CRA& m)
-    {
-        return solve(x, d, A, b, tag, m.iterationMethod());
-    }
-#endif
-
-#ifdef __LINBOX_HAVE_IML
-        //! IML wrapper.
-        //! @bug not recognised as template spec...
-    BlasVector<Givaro::ZRing<Integer>>&
-    solveIML(BlasVector<Givaro::ZRing<Integer>>& x, Givaro::ZRing<Integer>::Element & d,
-             const BlasMatrix<Givaro::ZRing<Integer>>& B, const BlasVector<Givaro::ZRing<Integer>>& b,
-             const Method::IML& m)
-    {
-        THIS_CODE_COMPILES_BUT_IS_NOT_TESTED; // NOT MUCH
-        switch (m.routine()) {
-            case(1) : { /*  non singular */
-                linbox_check(B.rowdim()==B.coldim());
-                mpz_t * mp_A = REINTERP_IML_CONST(B.getPointer()) ;
-                    // reinterpret_cast<mpz_t*>(const_cast<Givaro::ZRing<Integer>::Element*>((B.getPointer())));
-                    //B.getConstPointer() ?
-                mpz_t * mp_B = REINTERP_IML_CONST(&b[0]);
-                mpz_t * mp_N = REINTERP_IML(&x[0]);
-                mpz_t mp_D ;
-                mpz_init(mp_D);
-                if (!m.computeRNS()) {
-                    IML::nonsingSolvLlhsMM(IML::RightSolu,(long)B.rowdim(),1,
-                                           mp_A, mp_B, mp_N, mp_D);
-                }
-                else {
-                    long n = (long)B.coldim();
-                    long basislen = 1;
-                    mpz_t mp_alpha, mp_maxInter;
-                    IML::FiniteField qh;
-                    IML::FiniteField *basis, **basiscmb;
-                    IML::Double ** ARNS ;
-                    mpz_init(mp_alpha);
-                    IML::maxMagnMP(mp_A, n, n, n, mp_alpha);
-                    mpz_init_set_ui(mp_maxInter, 1);
-                    mpz_addmul_ui(mp_maxInter, mp_alpha, 2);
-                    qh            = IML::RNSbound(n);
-                    basiscmb      = IML::findRNS(qh, mp_maxInter, &basislen);
-                    basis = basiscmb[0];
-                    mpz_clear(mp_maxInter);
-                    mpz_clear(mp_alpha);
-                        /*  CRNS[i] = [A_11, A_12] mod basis[i] */
-                    ARNS = IML_XMALLOC(IML::Double *, (size_t)basislen);
-                    for (long i = 0; i < basislen; ++i)
-                    {
-                        ARNS[i] = IML_XMALLOC(IML::Double,(size_t)(n*n));
-                        for (long j = 0; j < n; ++j)
-                            for (long l = 0; l < n; l++)
-                                ARNS[i][j*n+l] = (IML::Double) mpz_fdiv_ui(mp_A[j * n + l], basis[i]);
-                    }
-                    IML::nonsingSolvRNSMM(IML::RightSolu, n, 1, basislen,
-                                          basis, ARNS, mp_B, mp_N, mp_D);
-
-                }
-                mpz_set(d.get_mpz(),mp_D);
-                mpz_clear(mp_D);
-
-            }
-                    //!@todo wrap nonsingSolvRNSMM too
-                break;
-            case (2) : { /*  certified */
-                mpz_t * mp_A = REINTERP_IML_CONST(B.getPointer()) ;
-                    // reinterpret_cast<mpz_t*>(const_cast<Givaro::ZRing<Integer>::Element*>((B.getPointer())));
-                    //B.getConstPointer() ?
-                mpz_t * mp_b = REINTERP_IML_CONST(&b[0]);
-                mpz_t * mp_N = REINTERP_IML(&x[0]);
-                mpz_t mp_D ;
-                mpz_init(mp_D);
-                mpz_t * mp_NZ = NULL;
-                mpz_t mp_DZ ;
-                if (m.certificate()) {
-                    mp_NZ = IML_XMALLOC(mpz_t,x.size());
-                    for (size_t i = 0; i < x.size(); ++i) { mpz_init(mp_NZ[i]); }
-                    mpz_init(mp_DZ);
-                }
-                if ( !m.reduced() ) {
-                    IML::certSolveMP(m.certificate(),
-                                     (long) B.rowdim(),(long)B.coldim(),
-                                     mp_A, mp_b, mp_N, mp_D,
-                                     mp_NZ, mp_DZ);
-                }
-                else {
-                    IML::certSolveRedMP(m.certificate(),
-                                        m.nullcol() //NULLSPACE_COLUMN
-                                        ,(long)B.rowdim(),(long)B.coldim(),
-                                        mp_A, mp_b, mp_N, mp_D,
-                                        mp_NZ, mp_DZ);
-
-                }
-                mpz_set(d.get_mpz(),mp_D);
-                mpz_clear(mp_D);
-            }
-                break;
-            default :
-                throw LinBoxError("unknownn routine from IML (choice 1/2). Got XXX");
-        }
-        return x;
-    }
-#endif
-
-    template <class RatVector, class Vector, class BB, class MyMethod>
-    RatVector& solve(RatVector& x, const BB& A, const Vector& b,
-                     const RingCategories::IntegerTag & tag,
-                     const MyMethod& M)
-    {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-
-        commentator().start ("Rational CRA Solve", "Rsolve");
-        typename BB::Field::Element den;
-        BlasVector<typename BB::Field> num(A.field(),A.coldim());
-        solve (num, den, A, b, tag, M);
-        typename RatVector::iterator it_x= x.begin();
-        typename BlasVector<typename BB::Field>::const_iterator it_num= num.begin();
-        integer n,d;
-        A.field().convert(d,den);
-        for (; it_x != x.end(); ++it_x, ++it_num){
-            A.field().convert(n, *it_num);
-            *it_x = typename RatVector::value_type(n, d);
-        }
-        commentator().stop ("done", NULL, "Rsolve");
-        return x;
-    }
-
-    template <class RatVector, class Vector, class BB, class MethodTraits>
-    RatVector& solve(RatVector& x, const BB& A, const Vector& b,
-                     const RingCategories::RationalTag & tag,
-                     const MethodTraits& m)
-    {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-        commentator().start ("Rational CRA Solve", "Rsolve");
-        // 0.7213475205 is an upper approximation of 1/(2log(2))
-        size_t bits = (size_t)(26 -(int)ceil(log((double)A.rowdim())*0.7213475205));
-        PrimeIterator<IteratorCategories::HeuristicTag> genprime( (unsigned) bits);
-        RationalCra< RationalCraBuilderEarlyMultip< Givaro::Modular<double> > > rra(3UL);
-        IntegerModularSolve<BB,Vector,MethodTraits > iteration(A, b, m);
-        Integer den;
-        Givaro::ZRing<Integer> Z ;
-        BlasVector<Givaro::ZRing<Integer>> num(Z,A.coldim());
-        rra(num, den, iteration, genprime);
-
-        auto&& it_x= x.begin();
-        for (auto it_num:num){
-            integer g = gcd( it_num, den);
-            *it_x = typename RatVector::value_type(it_num/g, den/g);
-            ++it_x;
-        }
-        commentator().stop ("done", NULL, "Rsolve");
-
-        return x;
-    }
-
-    template <class RatVector, class BB, class MethodTraits>
-    RatVector& solve(RatVector& x, const BB& A, const RatVector& b,
-                     const RingCategories::RationalTag & tag,
-                     const MethodTraits& m)
-    {
-        if ((A.coldim() != x.size()) || (A.rowdim() != b.size()))
-            throw LinboxError("LinBox ERROR: dimension of data are not compatible in system solving (solving impossible)");
-        commentator().start ("Rational CRA Solve", "Rsolve");
-        // 0.7213475205 is an upper approximation of 1/(2log(2))
-        size_t bits = (size_t)(26 -(int)ceil(log((double)A.rowdim())*0.7213475205));
-        PrimeIterator<IteratorCategories::HeuristicTag> genprime((unsigned) bits);
-        RationalCra< RationalCraBuilderEarlyMultip< Givaro::Modular<double> > > rra(3UL);
-        IntegerModularSolve<BB,RatVector,MethodTraits > iteration(A, b, m);
-        Integer den;
-        Givaro::ZRing<Integer> Z;
-        BlasVector<Givaro::ZRing<Integer>> num(Z,A.coldim());
-        rra(num, den, iteration, genprime);
-
-        auto&& it_x= x.begin();
-        for (auto it_num:num){
-            integer g = gcd( it_num, den);
-            *it_x = typename RatVector::value_type(it_num/g, den/g);
-            ++it_x;
-        }
-        commentator().stop ("done", NULL, "Rsolve");
-
-        return x;
-    }
-
-} // LinBox
-
-#include "linbox/config-blas.h"
-#ifdef __LINBOX_HAVE_CLAPACK
-#include "linbox/algorithms/numeric-solver-lapack.h"
-#include "linbox/algorithms/rational-solver-sn.h"
-namespace LinBox {
-    BlasVector<Givaro::ZRing<Integer>>&
-    solveNum(BlasVector<Givaro::ZRing<Integer>>& x, Givaro::ZRing<Integer>::Element & d,
-             const BlasMatrix<Givaro::ZRing<Integer>>& B, const BlasVector<Givaro::ZRing<Integer>>& b,
-             const Method::NumSymOverlap & m)
-    {
-        THIS_CODE_COMPILES_BUT_IS_NOT_TESTED; // NOT MUCH
-
-        typedef ParamFuzzy Field ;
-        typedef BlasMatrix<Field> FMatrix;
-
-        typedef LPS<FMatrix > NumSolver;
-        NumSolver numSolver;
-        bool e = false ;
-        Givaro::ZRing<Integer> Z;
-        RationalSolverSN<Givaro::ZRing<Integer>, NumSolver > rsolver(Z, numSolver, e);
-
-        int status = rsolver.solve(x, d, B, b);
-        if (status)
-            throw LinBoxError("fail") ;
-        return x;
-    }
-}
-#endif
-
-namespace LinBox {
-    BlasVector<Givaro::ZRing<Integer>>&
-    solveNum(BlasVector<Givaro::ZRing<Integer>>& x, Givaro::ZRing<Integer>::Element & d,
-             const BlasMatrix<Givaro::ZRing<Integer>>& B, const BlasVector<Givaro::ZRing<Integer>>& b,
-             const Method::NumSymNorm & m)
-    {
-            //THIS_CODE_COMPILES_BUT_IS_NOT_TESTED; // NOT MUCH
-
-        typedef Givaro::Modular<int32_t> ZField;
-            // typedef Givaro::Modular<double> ZField;
-        Givaro::ZRing<Integer> ZZ ;
-        RationalSolver<Givaro::ZRing<Integer>, ZField, PrimeIterator<IteratorCategories::HeuristicTag>, NumSymNormTraits> rsolver(ZZ);
-
-        int status = rsolver.solve(x, d, B, b);
-        if (status)
-                // throw "fail" ;
-            std::cerr << "fail:" << status << std::endl;
-        return x;
+        solveInPlace(xNum, xDen, A, b, Method::Auto());
     }
 }
 
-#endif // __LINBOX_solve_H
+//
+// solve
+//
 
-// Local Variables:
-// mode: C++
-// tab-width: 4
-// indent-tabs-mode: nil
-// c-basic-offset: 4
-// End:
-// vim:sts=4:sw=4:ts=4:et:sr:cino=>s,f0,{0,g0,(0,\:0,t0,+0,=s
+#include "./solve/solve-auto.h"
+
+// Elimination
+#include "./solve/solve-dense-elimination.h"
+#include "./solve/solve-elimination.h"
+#include "./solve/solve-sparse-elimination.h"
+
+// Integer-based
+#include "./solve/solve-cra.h"
+#include "./solve/solve-dixon.h"
+#include "./solve/solve-numeric-symbolic.h"
+
+// Blackbox
+#include "./solve/solve-blackbox.h"
+#include "./solve/solve-lanczos.h"
+#include "./solve/solve-wiedemann.h"
+
+#endif
